requirements.txt
asgiref==3.3.4
autopep8==1.5.7
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cloudinary==1.25.0
colorama==0.4.4
confusable-homoglyphs==3.2.0
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.45.0
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-heroku==0.3.1
django-registration==3.1.2
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.4
gunicorn==20.1.0
hvac==0.11.0
idna==3.2
loguru==0.5.3
oauthlib==3.1.1
psycopg2==2.8.5
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
validators==0.18.2
whitenoise==5.2.0
win32-setctime==1.0.3
